Update (and fix) date handling
Since the update to version 9 of NHS Frontend, dates are now easier to retrieve and set.

This means that the `items` array no longer needs to be set on the date inputs, and instead a `values` object can be set at the top level.

When retrieving dates from the routes, these can now be accessed using `data.batchExpiryDate.year` instead of `data['batchExpiryDate-year'].