Add errors and warnings based on estimated due date (#163)
This adds some error messages if the estimated due date is highly likely
to be an error, and some warnings if the estimated due date is outside
the recommended period, but still possible.

## Errors

### Blank

<img width="812" alt="Screenshot 2025-01-14 at 15 02 12"
src="https://github.com/user-attachments/assets/9afe8356-536a-4713-96fe-052e8d2f271c"
/>

### Due date is more than 37 weeks after vaccination date

This is an error as also highly unlikely to know they are pregnant at
this point

![Screenshot 2025-01-16 at 14 18
36](https://github.com/user-attachments/assets/558dc3f2-ed70-4301-a66c-ef8dadb570d8)

### Due date is more than 4 weeks before vaccination date

This is an error as almost no pregnancies will go beyond 44 weeks

<img width="908" alt="Screenshot 2025-01-14 at 14 58 25"
src="https://github.com/user-attachments/assets/868d0ffd-f4d1-4315-a589-aec5c3aca4c8"
/>

## Warnings

These are shown on a subsequent page, and users can choose to either
continue, go back (eg to fix a mistake) or cancel the vaccination (by
navigating away).

### RSV

This is only shown if the patient is not yet 28 weeks pregnant

![Screenshot 2025-01-16 at 14 22
29](https://github.com/user-attachments/assets/02398ba6-611b-4e73-bbd2-be5b9f585094)

### Pertussis

This is only shown if the patient is not yet 16 weeks pregnant

![Screenshot 2025-01-16 at 14 21
30](https://github.com/user-attachments/assets/38445572-e68b-409f-ac1a-a11e114f90b5)

---------

Co-authored-by: Anna-Sutton <184676072+Anna-Sutton@users.noreply.github.com>