Set page titles (#194)
This sets the `<title>` tag for all the pages in the new vaccination
journey. This is displayed on the browser Tab or window, as well as in
bookmarks, and is important for accessibility.

The format is:

`Page name - NHS Record a vaccination`

Generally they match the h1 but sometimes can be shorter, and should
avoid including patient details.