Warning for wrong age for this product (#238)
Adding another warning to the Record journey. This one is to warn that
the flu or covid product selected is not recommended for someone this
age.

The logic for when the equivalent live warning gets displayed is quite
complicated as it's based on age recommendations for each flu and covid
vaccine product. But it's not due to change for Spring 2025.

We will need to decide with devs, testers and clinical how granular to
go with dynamic content in this new version.

In the below example:
- in the heading, the product name is dynamic
- in the first sentence of text, name and age of patient are dynamic
- in the second sentence, name of product and who it's not recommended
for are dynamic. The most complex part is making who it's not
recommended for dynamic as there's lots of variation around exact age
windows in which certain vaccines are recommended.

So it might be simpler to keep the second sentence of text as.
'[Vaccine-product-name] is not recommended for someone this age.'

<img width="498" alt="image"
src="https://github.com/user-attachments/assets/11ea23b4-7c4f-4992-a283-2ba88a374279"
/>