Skip search results page if there’s a single result (#250)
This skips the search results page if there’s a single result and
instead goes straight to the patient details page.

If there are multiple results, or no results, then different pages are
shown, as before.

In future, if we are able to show the details of multiple matches, then
we could bring something like this back.

## Page removed

<img width="1042" alt="Screenshot 2025-03-11 at 09 27 19"
src="https://github.com/user-attachments/assets/23edc62d-5b43-4147-818b-d7c491cfced7"
/>