Updated text for 'Before you start' screen (#254)
Updated text. This version can be used for any user, irrespective of
permissions.

This example is for when a user tries to record when there is no-one at
their org with 'clinician' status.

There are 2 other variations of the wording on this screen: for when no
vax have been added, and for when no vax AND no clinician have been
added.