Change 'Appointments' to 'Lists' (#359)
The appointments designs are now known as lists.

The latest changes reflect work done to reduce the mental workload for
users when adding a 'list' and consist of 3 steps

- add a site or team
- add a date
- add nhs numbers

The list view page shows users in first name, last name order with nhs
number and date of birth for the recorder to check and links to 'Record'
or 'Remove' the patient from the list.

The home for 'lists' will update once lists are added and show lists by
tema or site at the first level, then ordered by date at the team/site
level down.

---------

Co-authored-by: Anna-Sutton <anna.sutton9@nhs.net>
Co-authored-by: Frankie Roberto <frankie@frankieroberto.com>