Change team to site on various pages (#409)
Changed 'team' to 'site' on various pages. 

This is because we have been using both terms in different parts of the
service to mean the same thing and we think it will be clearer for users
and for ourselves to be consistent. Data on the actual names of
sites/teams seems to align in most instances with a location so we felt
'site' is the more appropriate descriptor.