Upgrade NHS Prototype kit to version 7 (#453)
This upgrades the prototype to version 7 of the NHS Prototype Kit.

This is a manual process of copying the files over and merging any
changes - see
https://prototype-kit.service-manual.nhs.uk/how-tos/updating-the-kit

We had already update to NHS Frontend version 10 in #443, so this is
mainly copying over some improvements and bug fixes to the kit itself.