Update records section with filters (#468)
This is a design spike to look at how we could update the 'Records'
section so that, instead of just searching for records by patient name
or NHS number, you can also filter by:

* vaccination type
* batch number
* vaccinator

We’ve heard these could be useful in some circumstances, such as
correcting mistakes or checking that records were recorded correctly.

To accommodate these options, we could have a 'Filters' box, with the
most common filters shown by default, and more available under a
revealed section.

The records could be sorted by date and time recorded (most recent
first), with pagination at the bottom.

By default, all vaccination records (from your organisation) would be
shown (limited to most recent 500 or so?). This could mean that if you
just need to correct a mistake recorded within the past few minutes or
hour, you don't even need to use the filters at all, as it’ll be near
the top of the list.

I did also consider having the filters in a left hand column, but the
space is pretty tight given the number of filters and the number of
columns shown in the list. However it could be possible to make it work
if the content is reduced down a bit.

[Jira card](https://nhsd-jira.digital.nhs.uk/browse/RAVS-573)

## Screenshots

### Default filters

<img width="2240" height="1945" alt="localhost_3000_records"
src="https://github.com/user-attachments/assets/d5c0c613-5284-4451-89cb-f26b03bb29d8"
/>

### Additional filters

<img width="2240" height="3515" alt="localhost_3000_records (1)"
src="https://github.com/user-attachments/assets/49fa6e31-adb2-455a-9c11-8477eeaf986d"
/>