Use by organisation for multi org reporting (#570)
This updates the Home and Reports feature in "multi-organisation" mode,
so that it lists organisations instead of sites.

This does mean that in this mode, you wouldn't be able to narrow down
your report to only include Site A at Org 1 + Site B at Org 2 – however
we think this is less of a likely user need, and listing organisations
instead of sites should result in a simpler, shorter list.