Change order of questions (#525)
This moves the pages where the user searches for a patient and sees
their vaccination history to the beginning of the new Record
vaccinations journey.

This is because:

* some users like to use the patient history to assess eligibility for
the vaccine (for instance if the patient cannot remember)
* some users have a cohort of patients (for example in care homes) and
want to check if they are missing any vaccinations
* expectations from other systems are that the patient is found first

The original justification for having other questions first would be
that they support 'setting up' a clinic, answering the questions which
will apply to all vaccinations in that clinic first (such as location,
vaccinator and vaccine type). However there are enough ad-hoc
vaccinations and mixed clinic sessions that we now believe the benefits
of re-ordering the question outweigh this.

The repeat journey (same vax different patient, another vaccination for
a different patient) have also been changed a bit, putting the patient
finding first.

The co-admin journey is largely unchanged.